fixed issue. Also fixed some issues from the CI
