+++ condor_SUEP_WS.py
@@ -2,6 +2,7 @@ import argparse
 import os
 
 # Import coffea specific features
+from coffea import processor
 from coffea.processor import futures_executor, run_uproot_job
 
 # SUEP Repo Specific
