+++ workflows/SUEP_utils.py
@@ -474,3 +474,89 @@ def getTopTwoJets(self, tracks, indices, ak_inclusive_jets, ak_inclusive_cluster
         indices,
         (SUEP_cand, ISR_cand, SUEP_cluster_tracks, ISR_cluster_tracks),
     )
+
+def convert_coords(coords, tracks, nobj):
+    allowed_coords = ["cyl", "cart", "p4"]
+    if coords.lower() not in allowed_coords:
+        raise Exception(self.coords + " is not supported in GNN_convertEvents.")
+
+    if coords.lower() == "p4":
+        new_tracks = convert_p4(tracks, nobj)
+    elif coords.lower() == "cart":
+        new_tracks = convert_cart(tracks, nobj)
+    elif coords.lower() == "cyl":
+        new_tracks = convert_cyl(tracks, nobj)
+    else:
+        raise Exception()
+
+    return new_tracks
+
+
+def convert_p4(tracks, nobj=10):
+    """store objects in zero-padded numpy arrays"""
+
+    # bad naming, tracks dimensions is (events x tracks)
+    nentries = len(tracks)
+
+    l1Obj_p4 = np.zeros((nentries, nobj, 4))
+
+    px = to_np_array(tracks.px, maxN=nobj)
+    py = to_np_array(tracks.py, maxN=nobj)
+    pz = to_np_array(tracks.pz, maxN=nobj)
+    m = to_np_array(tracks.mass, maxN=nobj)
+
+    l1Obj_p4[:, :, 0] = px
+    l1Obj_p4[:, :, 1] = py
+    l1Obj_p4[:, :, 2] = pz
+    l1Obj_p4[:, :, 3] = m
+
+    return l1Obj_p4
+
+
+def convert_cyl(tracks, nobj=10):
+    """store objects in zero-padded numpy arrays"""
+
+    # bad naming, tracks dimensions is (events x tracks)
+    nentries = len(tracks)
+
+    l1Obj_cyl = np.zeros((nentries, nobj, 4))
+
+    pt = to_np_array(tracks.pt, maxN=nobj)
+    eta = to_np_array(tracks.eta, maxN=nobj)
+    phi = to_np_array(tracks.phi, maxN=nobj)
+    m = to_np_array(tracks.mass, maxN=nobj)
+
+    l1Obj_cyl[:, :, 0] = pt
+    l1Obj_cyl[:, :, 1] = eta
+    l1Obj_cyl[:, :, 2] = phi
+    l1Obj_cyl[:, :, 3] = m
+
+    return l1Obj_cyl
+
+
+def convert_cart(tracks, nobj=10):
+    """store objects in zero-padded numpy arrays"""
+
+    # bad naming, tracks dimensions is (events x tracks)
+    nentries = len(tracks)
+
+    l1Obj_cart = np.zeros((nentries, nobj, 4))
+
+    pt = to_np_array(tracks.pt, maxN=nobj)
+    eta = to_np_array(tracks.eta, maxN=nobj)
+    phi = to_np_array(tracks.phi, maxN=nobj)
+    m = to_np_array(tracks.mass, maxN=nobj)
+
+    l1Obj_cart[:, :, 0] = pt * np.cos(phi)
+    l1Obj_cart[:, :, 1] = pt * np.sin(phi)
+    l1Obj_cart[:, :, 2] = pt * np.sinh(eta)
+    l1Obj_cart[:, :, 3] = m
+
+    return l1Obj_cart
+
+
+def to_np_array(ak_array, maxN=100, pad=0):
+    """convert awkward array to regular numpy array"""
+    return ak.to_numpy(
+        ak.fill_none(ak.pad_none(ak_array, maxN, clip=True, axis=-1), pad)
+    )
